Quick fix to turn on logging.
pbsmrtpipe uses a different code-path than pypeflow in some places.
In this case, a new method uses the logger for a module, and
that logger is supposed to be set-up in main(). But main()
is part of pbcommand for a pbsmrtpipe task. So this hack will
use logging as configured by pbcommand for this particular
falcon module. I will set up a more general solution later.