Fix HGAP4 run_merge
The script uses a relative path for output, but it was being
run from the parent directory.

Similar to the fix for run_consensus (previous commit),
this fixes a step which has a different structure for
different sized input. That is a result of how HPC.daligner works.

re: SAT-482