Drop 'with cd'
These were added because of a bug in pbsmrtpipe long ago.
I just did some testing last night to discover that the SL
has fixed their bug. (I've no idea when.) So we no longer
need to change to the output directory ourselves.

I've added a check to catch regressions early in the workflow
(in build_rdb).